requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
astropy==5.3.2
Pillow==10.0.0
PyQt5==5.15.9
PyQtWebEngine==5.15.6
plotly==5.15.0
astroquery==0.4.6
requests==2.31.0
beautifulsoup4==4.12.2
astroalign==0.9.1
scipy==1.11.2
scikit-image==0.21.0
opencv-python==4.8.0.76